No Fixed a lot, went through and did a lot of optimizations.
Apparently, the progressbar on the splash screen was doing something like the fibonacci sequence, just to look busy. That's gone. Splash screen comes up right away. Marquee Progress animation.. Fixed the app opening in the background.

Fixed the Global Delete methods.
Changed tabSource, added a seperate ProgressBar for scanning directory.

This was an all nighter, and then some. But I think it was worthwhile.