Swapped from Google's Cloud Translate API to DeepL for machine-driven localization. Added first run of, German, French, Hungarian, Indonesian, Italian, Japanese, Ukrainian, and Chinese (Simplified).
Rewrote the way languages are handled to allow for new languages to just be dropped into the Program's global cultureInfos array.

First run prompt should ask which language you would like to use if the program doesn't detect that it's been translated into your locale.

Changed the way we load config.ini.. kind of not really. We load it once now. Statically, in Program.cs and it is publically available. This avoids possibilities for duplicate INI objects and resulting bugginess.

Other changes being worked on but not quite ready for live release.

Preparing to build, updating assembly information, 2.6.1.2 will be here soon™